A_P/requirements.txt
Flask
transformers
torch
PyMuPDF
Pillow
scikit-learn
